+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Changelog
 Notable changes to this project.
 
+## dev
+- signals: support upload to diagnostic tool
+
 ## [2.9.1] - 2021-09-27
 - add query `round_model_performances` (#60)
 - fix documentation